+++ test/intl/relative_time_format_test.exs
@@ -18,5 +18,18 @@ defmodule Intl.RelativeTimeFormatTest do
     test "in hours" do
       assert {:ok, "in 2 hours"} = Intl.RelativeTimeFormat.format(2, :hour, locale: :en)
     end
+
+    test "numeric always forces numeric output" do
+      assert {:ok, "1 day ago"} =
+               Intl.RelativeTimeFormat.format(-1, :day, locale: :en, numeric: :always)
+
+      assert {:ok, "in 1 day"} =
+               Intl.RelativeTimeFormat.format(1, :day, locale: :en, numeric: :always)
+    end
+
+    test "zero offset formats with the future pattern" do
+      assert {:ok, "in 0 days"} =
+               Intl.RelativeTimeFormat.format(0, :day, locale: :en, numeric: :always)
+    end
   end
 end